+++ packages/system/src/Clock/index.ts
@@ -1,3 +1,3 @@
-export * from "./impl"
+export { LiveClock } from "../Effect/runtime"
 export * from "./definition"
 export * from "./core"
